app/requirements.txt
Pillow
PyMuPDF
rarfile
